Rewire response spec generation to `ResponseSpec`
